+++ src/diagnostics.ts
@@ -21,7 +21,9 @@ import { outdatedEffectCodegen } from "./diagnostics/outdatedEffectCodegen.js"
 import { overriddenSchemaConstructor } from "./diagnostics/overriddenSchemaConstructor.js"
 import { returnEffectInGen } from "./diagnostics/returnEffectInGen.js"
 import { runEffectInsideEffect } from "./diagnostics/runEffectInsideEffect.js"
+import { schemaStructWithTag } from "./diagnostics/schemaStructWithTag.js"
 import { schemaUnionOfLiterals } from "./diagnostics/schemaUnionOfLiterals.js"
+import { schemaUnsafeTypes } from "./diagnostics/schemaUnsafeTypes.js"
 import { scopeInLayerEffect } from "./diagnostics/scopeInLayerEffect.js"
 import { strictBooleanExpressions } from "./diagnostics/strictBooleanExpressions.js"
 import { strictEffectProvide } from "./diagnostics/strictEffectProvide.js"
@@ -67,5 +69,7 @@ export const diagnostics = [
   strictEffectProvide,
   unknownInEffectCatch,
   runEffectInsideEffect,
-  schemaUnionOfLiterals
+  schemaUnionOfLiterals,
+  schemaStructWithTag,
+  schemaUnsafeTypes
 ]